requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.6
oauthlib==3.2.2
promise==2.3
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
urllib3==2.2.1
